requirements.txt
flask
firebase-admin
pyrebase4
python-dotenv
gunicorn